requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyJWT==2.8.0
python-decouple==3.8
django-cors-headers==4.3.1
django-ratelimit==4.1.0